Configures database initialization and entrypoint
Sets up the database with a prepared configuration and manages the container entrypoint.

This commit introduces a new `pg_hba.conf` file to configure client authentication, and copies it to the database data directory during initialization.
It also adds a `run_postgres.sh` script to manage the postgres process as a container entrypoint and creates the postgresql run directory if it doesn't exist.